+++ src/output.h
@@ -77,7 +77,7 @@ void blank_phylo_transmission_data_file(file_struct *);
 void write_phylo_transmission_data(file_struct *, char *, int);
 void write_phylo_individual_data(file_struct *, individual *,  long, int);
 void write_hivpos_individual_data(file_struct *, individual *, long, int);
-void print_partnership_network(file_struct *, char *, file_label_struct *, patch_struct *, int , int );
+void print_partnership_network(file_struct *, char *, file_label_struct *, patch_struct *, double , int );
 void print_partners_outside_community(char *, individual *, long , int , int );
 void initialise_partners_outside_community_file(char *, int );
 void print_assortativity(char *, debug_struct *, patch_struct *, int, file_struct *);
